+++ src/main/java/com/gudrhs8304/ticketory/dto/booking/InitBookingResponseDTO.java
@@ -1,5 +1,6 @@
 package com.gudrhs8304.ticketory.dto.booking;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 public record InitBookingResponseDTO(
@@ -9,5 +10,5 @@ public record InitBookingResponseDTO(
         String holdExpiresAt,       // ISO-8601 문자열
         String paymentStatus,       // "PENDING"
         String provider,
-        Integer totalPrice
+        BigDecimal totalPrice
 ) {}
